Added corgi control animations like stop and run randomly

+++ src/main/java/com/amh/common/CommonData.java
@@ -7,8 +7,8 @@ public class CommonData {
     public static final String GAME_TITLE = "DESKTOP PET";
     public static final Color BACKGROUND_COLOR = new Color(0, 0, 0, 0);
     public static final Dimension GAME_SCREEN_SIZE = Toolkit.getDefaultToolkit().getScreenSize();
-    public static final long FPS = 1_000_000_000 / 60;
-    public static final long UPS = 1_000_000_000 / 60;
+    public static final double FPS = 1_000_000_000 / 60;
+    public static final double UPS = 1_000_000_000 / 60;
 
     public static final String CORGI_WITH_TAIL_IMG ="corgi-asset-tail.png";
     public static final String CORGI_WITH_NO_TAIL_IMG ="corgi-asset-notail.png";
